Add CSV client import service
