+++ queues/workers/importClientsWorker.js
@@ -1,7 +1,8 @@
 const { Worker } = require('bullmq');
 const connection = require('../../services/redis');
+const { importClientsFromCsv } = require('../../services/importClients');
 
 module.exports = new Worker('import-clients', async job => {
-  // TODO: implement import logic
-  console.log(`Processing import job ${job.id}`, job.data);
+  const { filePath, enterpriseId } = job.data;
+  return await importClientsFromCsv(filePath, enterpriseId);
 }, { connection });
